This has the ability to send long strings of text because I've sorted out some of the flow control issues.
The main issue that remains is that I don't know how to deal with having
multiple transmit streams all sending at once.  I think I need to
separate the flow control for the sending of a single command with the
flow control with the sending a single transmitter's commands.  I just
don't know how to do that.