Convert the flow control for the whole comm into flow control for each transmitter, and some other changes.
In general, I've added an acknowledgement for every message sent to the
dongle.  This prevents overflow of the serial buffer, and then I've
added xon/xoff for each transmit operation queue.  It works with one
transmitter, but I don't have a good way to test for multiple
transmitters.  I also don't have support in the Python host program for
multiple transmitters.

The other changes include a change that uses the return value from
sprintf to pass to the serial send function.  I also reverted the
associated change that I applied for single-transmitter flow control.