Make a bunch of changes to the Arduino code because I have (sort-of) prototype hardware.
I've built a board around an Arduino Nano which has the LED's and keying circuit for one
transmitter.  It's mostly to test the design of the hardware, but it implied a bunch of
software changes.  For one thing, I've gone ahead and made it possible to remove the
serial support and support for a speed control.  I've also added support to light the
transmitter select LED.

I've also had to remove the support for the watchdog timer because it doesn't work with
the Nano.  Or it doesn't work with the cheap Chinese Nano clones.  Or something.  It
doesn't work on the board that controls my prototype, so I've removed it.