The arduino-mk package on my system has been updated.  This has required changes to my Makefile in order to make it work.  Functionality is a beautiful thing.
The BOARD_TAG is now "nano" because it's a nano, and there's this new "BOARD_SUB"
macro that must be atmega328old because the boards I'm using are these cheap
Chinese clones with the funky serial chip on it and if I use anything else, the
nano code starts running while avrdude is trying to update the flash.  Lastly,
the ARCHITECTURE has to be set to avr or nothing works.  This is because Arduino
no could mean AVR, or it could mean ARM, and I'm definitely using AVR.