Add the ability to select the active transmitter by clicking on the appropriate button on the display.  This is linked to a change in the arduino code.
The change to the keyer interface is pretty straightforward:  I add the code to send
the correct message to the keyer, and then I call that code when the transmitter is
selected.  That means that I have to pass the communications object to those buttons,
but that's no hill for a stepper.

The problem comes from the fact that the transmitter is selected at startup, and at
startup the arduino hasn't finished booting.  So, the command to select the active
transmitter is sent, but the UI then sits and waits for an acknowledgement that is
never going to arrive.  The fix is to send an acknowledgement, which really is more
of a "ready for a message" indicator than an acknowledgement, as soon as the serial
port is readon on the arduino.  This makes it all work.