Swap the PTT and key down lines
This keyer is designed such that it should be able to plug an aux cable in to
one of the key out ports and into a Hermes Lite 2 and have the radio key up
properly.  However, somewhere along the way I got something swapped, so I'm
fixing that by swapping it in the configuration.

There was supposed to be some sort of mnemonic for red and yellow meaning key
down and transmit, but since I can't remember what it was it must not have
been much of a mnemonic.

I also removed the include of the serial log header file.  I have no idea what
it was doing there.  It certainly isn't necessary and it prevented building.